% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imhen.R
\docType{package}
\name{imhen}
\alias{imhen}
\alias{imhen-package}
\title{imhen: Meteorological data for Vietnam.}
\description{
The imhen package provides monthly times series for 7 climatic variables in
67 climatic stations of Vietnam from 1960 to 2015.
}

