% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meteo}
\alias{meteo}
\title{Meteorological Vietnamese Data}
\format{A data frame with 36,660 rows and 10 variables:
\itemize{
  \item \code{year}: year of collection
  \item \code{month}: month of collection (ordered factor)
  \item \code{station}: name of the climatic station
  \item \code{Tx}: monthly average of the daily maximal temperatures
  (in centigrad)
  \item \code{Ta}: monthly average of the daily average temperatures
  (in centigrad)
  \item \code{Tm}: monthly average of the daily minimal temperatures
  (in centigrad)
  \item \code{aH}: monthly average of absolute humidity (in g/m3)
  \item \code{rH}: monthly average of relative humidity (in \%)
  \item \code{Rf}: monthly cumulative rainfall (in mm)
  \item \code{Sh}: monthly cumulative number of hours of sunshine
}}
\source{
Thai P.Q. et al. (2015) Seasonality of absolute humidity explains
seasonality of influenza-like illness in Vietnam. \emph{Epidemics}
\strong{13}: 65-73.
\href{http://marcchoisy.free.fr/pdf/Epidemics2015.pdf}{[PDF]}
}
\usage{
meteo
}
\description{
A dataset containing the meteorological data of Vietnam collected per month
from 69 climatic stations, from January 1960 to December 2015.
}
\details{
The variable \code{station} is a key shared with dataset
\code{\link{stations}}.
}
\examples{
## Extracting the meteorological data for the stations above an elevation of
## 500 meters:
sel <- subset(stations,elevation>500,station)
subset(meteo,station\%in\%sel)
}
\seealso{
The \code{\link{stations}} dataset that shares the variable
\code{station} with \code{meteo}.
}
\keyword{datasets}

