% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stations}
\alias{stations}
\title{Location and elevation of climatic stations}
\format{A Simple feature collection with 67 rows and 3 variables:
\itemize{
  \item \code{station}: name of the climatic station
  \item \code{elevation}: elevation of the climatic station (in m)
  \item \code{latitude}: latitude of the climatic station (in decimal
  coordinates)

}}
\source{
Thai P.Q. et al. (2015) Seasonality of absolute humidity explains
seasonality of influenza-like illness in Vietnam. \emph{Epidemics}
\strong{13}: 65-73.
\href{http://marcchoisy.free.fr/pdf/Epidemics2015.pdf}{[PDF]}
}
\usage{
stations
}
\description{
A dataset containing the coordinates and elevation of the climatic stations
of Vietnam.
}
\details{
The variable \code{station} is a key shared with dataset
\code{\link{meteo}}.
}
\examples{
## Extracting the meteorological data for the stations above an elevation of
## 500 meters:
sel <- subset(stations, elevation > 500, station)
subset(meteo, stations$station \%in\% sel$station)
}
\seealso{
The \code{\link{meteo}} dataset that shares the variable
\code{station} with \code{stations}.
}
\keyword{datasets}
