% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteoVN.R
\docType{package}
\name{meteoVN}
\alias{meteoVN}
\alias{meteoVN-package}
\title{meteoVN: Meteorological data for Vietnam.}
\description{
The meteoVN package provides monthly times series for 7 climatic variables in
67 climatic stations of Vietnam from 1960 to 2015.
}

