% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meteo}
\alias{meteo}
\alias{meteo_r}
\alias{meteo_intagg_2008_2017}
\title{Meteorological Vietnamese Data}
\format{A data frame with 37,848 rows and 10 variables:
\itemize{
  \item \code{year}: year of collection
  \item \code{month}: month of collection (ordered factor)
  \item \code{station} or \code{province}: name of the climatic station (for
  \code{meteo} and \code{meteo_r}) or the province (for /code {meteo_intagg_2008_2017})
  \item \code{Ta}: monthly average of the daily average temperatures
  (in centigrad)
  \item \code{Tx}: monthly average of the daily maximal temperatures
  (in centigrad)
  \item \code{Tm}: monthly average of the daily minimal temperatures
  (in centigrad)
  \item \code{Rf}: monthly cumulative rainfall (in mm)
  \item \code{aH}: monthly average of absolute humidity (in g/m3)
  \item \code{rH}: monthly average of relative humidity (in \%)
  \item \code{Sh}: monthly cumulative number of hours of sunshine
}}
\source{
Thai P.Q. et al. (2015) Seasonality of absolute humidity explains
seasonality of influenza-like illness in Vietnam. \emph{Epidemics}
\strong{13}: 65-73.
\href{http://marcchoisy.free.fr/pdf/Epidemics2015.pdf}{[PDF]}
}
\usage{
meteo

meteo_r

meteo_intagg_2008_2017
}
\description{
A dataset containing the meteorological data of Vietnam collected per month
from 69 climatic stations, from January 1960 to December 2017. Two versions
exist in this package, \code{meteo_r} is the tidy version of the raw data and
\code{meteo} is the tidy, clean version of the raw data, with the temperatures
were checked and adjusted if necessary. Additionnally, aggregates of these
climatic variables by the province definition of the most recent period
(starting in 2008) are avaiable from 2008 to 2017 in \code{meteo_intagg_2008_2017}.
}
\details{
The variable \code{station} is a key shared with dataset \code{\link{stations}}.
The variable \code{province} is a key shared with the most recent polygon
definitions of the provinces of Vietnam \code{gadmVN::gadm()$province}. The
aggregated data are generated by interpolation and aggregation weighted by the
local population density, using the 2009 data from
\href{http://www.worldpop.org.uk}{WorldPop}. The method used is detailed
\href{https://choisy.github.io/2018/11/08/computing-weather-variables-by-province/}{here}.
}
\examples{
## Extracting the meteorological data for the stations above an elevation of
## 500 meters:
sel <- subset(stations, elevation > 500, station)
subset(meteo, stations$station \%in\% sel$statiom)
}
\seealso{
The \code{\link{stations}} dataset that shares the variable
\code{station} with \code{meteo}.
}
\keyword{datasets}
